requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
openai==1.99.5
faster-whisper==1.1.1
requests==2.32.5
rumps==0.4.0
pynput==1.8.1
